Hello Splotch developers,
my name is Klaus Reuter and I'm a member of the application support group at
the Garching computing centre (RZG).

I'm contributing a reader for Gadget HDF5 files which can be accessed via
"simtype=8" in the parameter file.
In addition, I added an entry to the Makefile for the visualization cluster at
RZG.

Cheers,
Klaus (khr@rzg.mpg.de)


git-svn-id: http://www.mpa-garching.mpg.de/svn/cosmo-group/kdolag/splotch@14416 019ef673-9300-0410-806d-f16013bd1951